requirements.txt
annotated-types==0.7.0
anyio==4.7.0
cachetools==5.5.0
camelot-py==0.11.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
et_xmlfile==2.0.0
fastapi==0.115.6
fonttools==4.55.3
ghostscript==0.7
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.155.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httplib2==0.22.0
idna==3.10
kiwisolver==1.4.7
matplotlib==3.9.4
numpy==2.2.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20240706
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
pypdf==5.1.0
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
starlette==0.41.3
tabulate==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wordcloud==1.9.4